remove DefaultAuthPlugin in UserSpec and move EncodedPassword to executor
